feat(organize): expose complete dated topic cards
