+++ src/yantrikdb/__init__.py
@@ -22,11 +22,13 @@ from yantrikdb.organize import (
     OrganizationPlan,
     TopicHandle,
     assign_evidence_to_handles,
+    load_persisted_topic_cards,
     organize_evidence,
     organize_concerns,
     persist_concerns,
     persist_organization,
     recall_organized,
+    topic_card_document,
     validate_topic_handles,
     validate_concern_items,
 )
@@ -60,11 +62,13 @@ __all__ = [
     "OrganizationPlan",
     "validate_topic_handles",
     "assign_evidence_to_handles",
+    "load_persisted_topic_cards",
     "persist_organization",
     "organize_evidence",
     "organize_concerns",
     "persist_concerns",
     "recall_organized",
+    "topic_card_document",
     "validate_concern_items",
     "check_all_triggers",
     "recall_reranked",
